[CODE] Move get_browser and default_browser to utils
